+++ src/main/java/com/cardgarden/project/controller/card/like/CardLikeController.java
@@ -1,4 +1,4 @@
-package com.cardgarden.project.controller;
+package com.cardgarden.project.controller.card.like;
 
 import java.util.HashMap;
 import java.util.Map;
