+++ src/main/webapp/WEB-INF/views/recommend/ai.jsp
@@ -13,28 +13,33 @@
 		padding: 0;
 		margin: 0;
 	}
+	
 	.pattern-wrap {
+		display:flex;
+		flex-direction: column;
 	    width: 100%;
 	    min-height: 100vh;
 	    background-color: var(--main);
 	    display: flex;
-	    justify-content: center;
-	    align-items: flex-start;
-	    padding-top: 60px;
+	    ustify-content: flex-start;
+	    /* align-items: flex-start; */
+	    align-items: center;
+	    padding-top: 0px;
 	}
 
 	.pattern-container {
-	    display: flex;
-	    flex-wrap: wrap;
-	    gap: 24px;
-	    max-width: 1200px;
-	    width: 100%;
-	    margin: auto;
-	    background: none;
-	    border: none;
-	    box-shadow: none;
-	    padding: 0;
-	    text-align : center;
+	  display: grid;
+	  grid-template-columns: repeat(4, 1fr);
+	  gap: 24px;
+	  max-width: 1200px;
+	  width: 100%;
+	  margin: auto;
+	  background: white;
+	  border: none;
+	  box-shadow: none;
+	  padding: 25px;
+	  text-align: center;
+	  min-height: 370px;
 	}
 	.pattern-title-main {
 	    font-size: 28px;
@@ -45,23 +50,37 @@
 	}
 	
 	.pattern-group {
-	    flex: 1 1 270px;
-	    max-width: 300px;
-	    min-width: 220px;
-	    margin-bottom: 24px;
-	    padding: 22px;
-	    border: 1.5px solid var(--m2);
-	    border-radius: 12px;
-	    box-shadow: 0 1px 6px rgba(143,176,152,0.09);
-	    background: #fff;
-	    display: flex;
-	    flex-direction: column;
-	    gap: 7px;
+	  position: relative;
+	  cursor: pointer;
+	  transition:
+	    border-color 0.22s cubic-bezier(0.4,0,0.2,1),
+	    box-shadow 0.22s cubic-bezier(0.4,0,0.2,1),
+	    background 0.22s cubic-bezier(0.4,0,0.2,1);
+	  border: 1.5px solid var(--m2);
+	  border-radius: 16px;
+	  box-shadow: 0 2px 10px rgba(143,176,152,0.07);
+	  background: #fafcfb;  /* 기본 연한 회녹색 계열 배경 */
+	  padding: 22px;
+	  display: flex;
+	  flex-direction: column;
+	  gap: 10px;
+	  min-width: 220px;
+	  max-width: 100%;
 	}
+	/* 호버(마우스 올릴 때) 스타일 */
 	.pattern-group:hover {
-	    box-shadow: 0 3px 14px rgba(143,176,152,0.18);
-	    background: #fff;
-	    border-color: var(--m1);
+	  border-color: var(--m1);
+	  background: #F8FBF8;    /* 연초록~연주황 섞인 듯한 느낌 */
+	  box-shadow: 0 6px 24px 0 rgba(143,176,152,0.14);
+	}
+	/* 선택된 카드 (radio:checked) 스타일 */
+	.pattern-group.selected {
+	  border-color: var(--m1) !important;
+	  background: #F1F5EB;
+	  box-shadow:
+	    0 0 0 0.4px var(--m1),
+	    0 6px 28px 0 rgba(255,184,74,0.12),
+	    0 3px 18px rgba(143,176,152,0.16);
 	}
 	.pattern-title {
 	    font-weight: bold;
@@ -82,7 +101,7 @@
 	    padding: 8px 0 8px 18px;
 	    border-left: 4px solid var(--m1);
 	    margin-bottom: 7px;
-	    background: var(--main);
+	    background: white;
 	    border-radius: 5px;
 	    font-size: 15px;
 	    color: var(--m3);
@@ -136,7 +155,39 @@
 		/* margin-bottom: 22px */
 		text-align: center;	
 	}
+	.btn-all {
+	  display: flex;
+	  flex-direction: row;
+	  justify-content: flex-start;
+	  align-items: center;
+	  gap: 14px;
+	  margin-bottom: 32px; /* 버튼 아래에 여백 */
+	  margin-left: 24px;
+	}
+
+	.pattern-list{
+		background: white;
+	}
+	/* .ai-form{
+		display: flex;
+		
+		justify-content: center;
+		height: 100vh;
+	} */
+	.pattern-radio {
+	  position: absolute;
+	  opacity: 0;
+	  pointer-events: none;
+	  width: 0; height: 0;
+	  margin: 0; padding: 0;
+	}
+	/* 선택된 패턴 스타일 */
+	.pattern-radio:checked + .pattern-title,
+	.pattern-radio:checked ~ .benefit-row {
+	  /* 이건 title, benefit-row에 주는 방법이지만... */
+	}
 </style>
+
 
 <script>
   function checkOnlyOne(element) {
@@ -161,47 +212,63 @@
  </div>
 <!-- 로딩화면 끝 -->
 	<div class="pattern-wrap">
-	  <c:choose>
-	    <c:when test="${empty patternList}">
-	      <div class="pattern-container" style="display:block;">
-	        <p class="non-pattern">등록된 소비패턴이 없습니다.</p>
-	        <form action="${cpath}/ConsumptionPattern/inCon" method="get">
-	          <button type="submit" class="button-primary">소비패턴 입력하러 가기</button>
-	      	</form>
-	      </div>
-	      
-	    </c:when>
-	    <c:otherwise>
-	      <form method="post" action="${cpath}/recommend/selectPattern" onsubmit="event.preventDefault(); showMaskAndSubmit(this);">
-	        <div class="pattern-container">
-	          <c:forEach var="entry" items="${patternList}">
-	            <div class="pattern-group" onclick="document.getElementById('pattern_${entry.key}').click();">
-	              <div class="pattern-title">
-	               <input type="radio" class="pattern-radio" name="patternId"
-					       value="${entry.key}" id="pattern_${entry.key}">
-					<c:if test="${not empty entry.value}">
-					    <label for="pattern_${entry.key}">
-					        ${entry.value[0].pattern.pattern_name}
-					    </label>
-					</c:if>
+  <c:choose>
+    <c:when test="${empty patternList}">
+      <div class="pattern-container" style="display:block;">
+        <p class="non-pattern">등록된 소비패턴이 없습니다.</p>
+        <form action="${cpath}/ConsumptionPattern/inCon" method="get">
+          <button type="submit" class="button-primary">소비패턴 입력하러 가기</button>
+        </form>
+      </div>
+    </c:when>
+    <c:otherwise>
+      <form method="post" action="${cpath}/recommend/selectPattern" onsubmit="event.preventDefault(); showMaskAndSubmit(this);" class="ai-form">
+        <!-- 버튼: 상단에 한 줄 -->
+        <div class="btn-all">
+          <button type="submit" class="button-primary" style="margin-right: 12px;">
+            선택한 패턴으로 카드 추천받기
+          </button>
+          <a class="button-primary" href="${cpath}/ConsumptionPattern/inCon">
+            소비패턴 입력하러 가기
+          </a>
+        </div>
+        <!-- 소비패턴 카드들: 버튼 아래 -->
+        <div class="pattern-container">
+		  <c:forEach var="entry" items="${patternList}">
+		    <label class="pattern-group" for="pattern_${entry.key}">
+		      <input type="radio" class="pattern-radio" name="patternId"
+		             value="${entry.key}" id="pattern_${entry.key}" autocomplete="off">
+		      <div class="pattern-title">
+		        <c:if test="${not empty entry.value}">
+		          ${entry.value[0].pattern.pattern_name}
+		        </c:if>
+		      </div>
+		      <c:forEach var="dto" items="${entry.value}">
+		        <div class="benefit-row">
+		          ${dto.category.benefitCategory_name} : <b>${dto.detail.amount}</b> 원
+		        </div>
+		      </c:forEach>
+		    </label>
+		  </c:forEach>
+		</div>
 
-	              </div>
-	              <c:forEach var="dto" items="${entry.value}">
-	                <div class="benefit-row">
-	                  ${dto.category.benefitCategory_name} : <b>${dto.detail.amount}</b> 원
-	                </div>
-	              </c:forEach>
-	            </div>
-	          </c:forEach>
-	        </div>
-	        <button type="submit" class="button-primary" style="margin: 40px auto 0 auto; display: block;">
-	          선택한 패턴으로 카드 추천받기
-	        </button>
-	         <a class="button-primary" href="${cpath}/ConsumptionPattern/inCon">소비패턴 입력하러 가기</a>
-	      </form>
-	    </c:otherwise>
-	  </c:choose>
-	</div>
+      </form>
+      <script type="text/javascript">
+		document.querySelectorAll('.pattern-radio').forEach(function(radio) {
+		  radio.addEventListener('change', function() {
+		    document.querySelectorAll('.pattern-group').forEach(function(label){
+		      label.classList.remove('selected');
+		    });
+		    if (radio.checked) {
+		      radio.closest('.pattern-group').classList.add('selected');
+		    }
+		  });
+		});
+	</script>
+    </c:otherwise>
+  </c:choose>
+</div>
+
